Likely Plan B
Due to significant technical difficulties, it is likely that milestone 2 will be completed with only the arduino.  This code allows the arduino to compute all the logic and control by itself.  The only implementation required for full operation is turnLeft() and turnRight() which is waiting for Robert.  But first we have to get the steering to fully function before.